add update Item validation helper and update parent component
